Added switch to disable local ray temp

+++ nataili/util/voodoo.py
@@ -8,15 +8,19 @@ from typing import Dict, List, Tuple, TypeVar
 import torch
 
 import ray
+from nataili import disable_local_ray_temp
 from nataili.util import logger
 
 T = TypeVar("T")
 
-ray_temp_dir = "./ray"
-shutil.rmtree(ray_temp_dir, ignore_errors=True)
-os.makedirs(ray_temp_dir, exist_ok=True)
-ray.init(_temp_dir=ray_temp_dir)
-logger.init(f"Ray temp dir '{ray_temp_dir}'", status="Prepared")
+if not disable_local_ray_temp.active:
+    ray_temp_dir = "./ray"
+    shutil.rmtree(ray_temp_dir, ignore_errors=True)
+    os.makedirs(ray_temp_dir, exist_ok=True)
+    ray.init(_temp_dir=ray_temp_dir)
+    logger.init(f"Ray temp dir '{ray_temp_dir}'", status="Prepared")
+else:
+    logger.init_warn(f"Ray temp dir'", status="OS Default")
 
 
 def performance(f: T) -> T:
